ExtraP Bug-fix in No Metadata Case
Fix a bug in the case where the metadata had to be given as a list when the parameters were not provided for Extra-P. This method assumed that the files would be input in a certain manner. This patch makes the behavior more robust as it requires a mapping from profile to parameter so this bug will not happen.

* Update unit test

* Formatting

* Fix bug when profiles are read in an unexpected order